feat:profile:geo: Swagger  refactor:profile:geo: reorganize classes into packages  fix:security: Swagger
