+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/management/controller/AdminDataImportController.java
@@ -1,16 +1,17 @@
-package ru.pavbatol.myplace.geo.management;
+package ru.pavbatol.myplace.geo.management.controller;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;
+import ru.pavbatol.myplace.geo.management.service.DataImportService;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -19,15 +20,18 @@ import java.time.format.DateTimeFormatter;
 @RestController
 @RequestMapping("admin/geo/upload") // TODO: Add prefix for APi
 @RequiredArgsConstructor
-public class DataImportController {
+@Tag(name = "Admin: Import geo data", description = "API for importing geo data from CSV")
+public class AdminDataImportController {
     private final DataImportService dataImportService;
 
     @Value("${spring.application.name}")
     private String serviceName;
 
     @PostMapping("/csv")
-    public ResponseEntity<StreamingResponseBody> uploadCsv(@RequestParam("file") MultipartFile file,
-                                                           @RequestParam(value = "responseExportWithId", defaultValue = "false") boolean responseExportWithId) { // TODO: Add this into html form
+    @Operation(summary = "import data", description = "importing geo data from CSV file")
+    public ResponseEntity<StreamingResponseBody> uploadCsv(
+            @Parameter(description = "CSV file to upload", required = true) @RequestPart("file") MultipartFile file,
+            @RequestParam(value = "responseExportWithId", defaultValue = "false") boolean responseExportWithId) {
         log.debug("POST uploadCsv with file sized {} byte; responseExportWithId={}", file.getSize(), responseExportWithId);
         long maxFileSizeMb = 5;
